Mention the potential need for port forwarding.
OSX users (or anyone else running docker in a VM) may not know that there is an additional step to access the docker UI from their browser
Users new to docker may not even yet appreciate the distinction between the docker OS/container interations and the VM host machine / VM guest machine interactions. It's worth calling this out in the Quickstart, and giving an extended description in an FAQ page.